--- backend/src/main/java/dao/BBSDAO.java
@@ -1,16 +1,0 @@
-package dao;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.springframework.stereotype.Repository;
-
-import domain.BBS;
-
-@Mapper
-@Repository
-public interface BBSDAO {
-	
-	List<BBS> getBBSPageList();
-
-}
