--- backend/src/main/java/domain/Map.java
@@ -1,10 +1,0 @@
-package domain;
-
-
-//자료구조 연습장
-//public class Map {
-
-//	List Map<Long, Post> = new HashMap<>();
-//	postMap.put(1L, new Post("title", "content"))
-//	Post post = postMap.get(1L)
-//}
