--- backend/src/main/java/domain/Member.java
@@ -1,15 +1,0 @@
-package domain;
-
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class Member {
-	private String id;
-	private String pwd;
-	private String name;
-	private String email;
-	
-}
