+++ backend/src/main/java/jay/toy/BackendApplication.java
@@ -4,10 +4,11 @@ import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import domain.Student;
 
-@MapperScan("dao")
+
+@MapperScan("jay.toy.dao")
 @SpringBootApplication
+//(scanBasePackages = {"jay.toy.dao"})
 public class BackendApplication {
 
 	public static void main(String[] args) {
