+++ backend/src/main/java/jay/toy/domain/Map.java
@@ -1,0 +1,10 @@
+package jay.toy.domain;
+
+
+//자료구조 연습장
+//public class Map {
+
+//	List Map<Long, Post> = new HashMap<>();
+//	postMap.put(1L, new Post("title", "content"))
+//	Post post = postMap.get(1L)
+//}
