+++ backend/src/main/java/jay/toy/domain/Member.java
@@ -1,0 +1,15 @@
+package jay.toy.domain;
+
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Member {
+	private String id;
+	private String pwd;
+	private String name;
+	private String email;
+	
+}
