+++ backend/src/main/java/jay/toy/domain/MyClass.java
@@ -1,0 +1,28 @@
+package jay.toy.domain;
+
+import lombok.Data;
+
+
+//학습용 - 지워질 것
+//@Data
+//public class MyClass {
+//	private int value;
+//	private String name;
+//	
+//
+//	
+//	
+//}
+//public class Main{
+//	public static void main(String[] args) {
+//		MyClass obj = new MyClass();
+//		obj.setValue(10);
+//		obj.setName("Example");
+//		System.out.println(obj.toString());
+//	}
+//}
+
+
+//
+//이 CreateBbsRequest DTO는 컨트롤러에서 글쓰기 요청을 처리할 때 사용되며, @RequestBody 애노테이션과 함께 사용되어 JSON 형태로 넘어온 데이터를 이 클래스의 인스턴스로 변환하게 됩니다.
+//이 말이 이해가 안가
