+++ backend/src/main/java/jay/toy/dto/request/BBSListRequest.java
@@ -1,0 +1,19 @@
+package jay.toy.dto.request;
+
+
+import java.time.LocalDateTime;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class BBSListRequest {
+	
+	private int seq;
+	private String title;
+	private LocalDateTime createdAt;
+	
+}
