+++ backend/src/main/java/jay/toy/dto/request/CreateBBSRequest.java
@@ -1,0 +1,17 @@
+package jay.toy.dto.request;
+
+import java.time.LocalDateTime;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreateBBSRequest {
+	
+	private String title;
+	private String contents;
+	
+}
