+++ backend/src/main/java/jay/toy/dto/response/CreateBBSResponse.java
@@ -1,0 +1,14 @@
+package jay.toy.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreateBBSResponse {
+	
+	 
+
+}
