+++ backend/src/main/resources/application.properties
@@ -4,6 +4,9 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.username=toy
 spring.datasource.password=pass
 
+mybatis.mapper-locations=classpath:sqls/*.xml
+
+
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=create
 spring.jpa.properties.hibernate.format_sql=true;
