+++ backend/src/main/resources/sqls/Bbs.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="dao.BBSDAO">
+<mapper namespace="jay.toy.dao.BBSDAO">
  	<select id="now" resultType="string">
  		select now();
  	</select>
- 	<select id="getBBSPageList" resultType="domain.BBS">
+ 	<select id="getBBSPageList" resultType="jay.toy.domain.BBS">
  	select * from bbs
  	</select>
 </mapper>
